requirements_versions.txt
accelerate==0.13.2
albumentations==0.4.3
basicsr==1.4.2
clean_fid==0.1.34
einops==0.3.0
facexlib==0.2.5
fonts~=0.0.3
font-roboto
imageio==2.9.0
jsonmerge==1.8.0
kornia==0.6.0
lark==1.1.2
matplotlib==3.6.1
numexpr==2.8.4
omegaconf==2.1.1
opencv_python_headless==4.6.0.66
packaging==21.3
pandas==1.5.1
piexif==1.1.3
pip==22.2.2
psutil==5.9.3
pudb==2019.2
Pygments==2.13.0
pytorch_lightning==1.4.2
pytz==2022.5
regex==2022.9.13
resize_right==0.0.2
scikit_image==0.19.3
setuptools==65.4.0
timm==0.6.11
tqdm==4.64.1
wandb==0.13.4
pyside6~=6.4.0.1
numpy~=1.23.1
pillow~=9.2.0
pyyaml~=6.0
opencv-python~=4.5.5.64
clip
scipy~=1.9.3
ftfy~=6.1.1
future~=0.18.2
gfpgan~=1.3.5
realesrgan~=0.3.0
scikit-image~=0.19.3
requests~=2.25.1
gitdb~=4.0.9
typing_extensions~=4.3.0
cffi~=1.15.1
attrs~=22.1.0
wheel~=0.37.1
cryptography~=38.0.1
rsa~=4.9
pyasn1~=0.4.8
wcwidth~=0.2.5
six~=1.16.0
atomicwrites~=1.4.0
pyinstaller~=5.6.2
lmdb~=1.3.0
urwid~=2.1.2
colorama~=0.4.6
yapf~=0.32.0
idna~=2.10
multidict~=6.0.2
filelock~=3.8.0
smmap~=5.0.0
lpips~=0.1.4
numba~=0.56.3
llvmlite~=0.39.1
sklearn~=0.0
scikit-learn~=1.1.3
fsspec~=2022.10.0
tensorboard~=2.10.1
urllib3~=1.26.12
click~=8.1.3
shortuuid~=1.0.9
boto3~=1.26.3
pathtools~=0.1.2
gevent~=22.10.2
promise~=2.3
joblib~=1.2.0
addict~=2.4.0
networkx~=2.8.7
aiohttp~=3.8.3
yarl~=1.8.1
cachetools~=5.2.0
certifi~=2022.9.24
imgaug~=0.2.6
markupsafe~=2.1.1
python-dateutil~=2.8.2
pyarrow~=7.0.0
aiosignal~=1.2.0
frozenlist~=1.3.1
chardet~=4.0.0
tifffile~=2022.10.10
shiboken6~=6.4.0.1
filterpy~=1.4.5
markdown~=3.4.1
oauthlib~=3.2.2
win_inet_pton~=1.1.0
werkzeug~=2.2.2
contourpy~=1.0.6
fonttools~=4.38.0
jsonschema~=4.16.0
pyparsing~=3.0.9
pyrsistent~=0.18.1
cycler~=0.11.0
kiwisolver~=1.4.4
botocore~=1.29.3
bottle~=0.12.23
tokenizers~=0.12.1
pydantic~=1.10.2
fairscale~=0.4.12
k-samplers~=1.0.0
open_clip_torch~=2.0.2
torchsde
fastapi
uvicorn
pydantic
transformers~=4.24.0
diffusers~=0.9.0
safetensors~=0.2.5
